implement custom shaders for the OpenGLES platform (also replace most of its code with code from OpenGL3)

+++ Wrappers/MyGUI_Export/MyGUI_ExportRenderManager.cpp
@@ -87,7 +87,7 @@ namespace MyGUI
 		unsigned long now_time = timer.getMilliseconds();
 		unsigned long time = now_time - last_time;
 
-		onFrameEvent((float)((double)(time) / (double)1000));
+		onFrameEvent(time / 1000.0f);
 
 		last_time = now_time;
